mosguito/requirements.txt
Django==3.2.15
django_filter==22.1
djangorestframework==3.13.1
django_rql==4.2.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
habanero==1.2.2
pandas==1.3.5
requests==2.28.1
six==1.16.0
# mysqlclient==2.1.1
gunicorn==20.1.0
whitenoise==6.3.0
celery
redis
